requirements.txt
fastapi
uvicorn[standard]
pydantic
psycopg[binary]
PyYAML
python-dotenv
